# openAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Freya Salon Backend API",
    "description": "\n    ## Freya Beauty Salon Management System API\n    \n    Bu API Freya go'zallik saloni boshqaruv tizimi uchun yaratilgan.\n    \n    ### Asosiy funksiyalar:\n    - 🔐 **Autentifikatsiya va avtorizatsiya** (JWT token)\n    - 👥 **Foydalanuvchilar boshqaruvi** (Admin, User, Employee)\n    - 🏢 **Salon boshqaruvi** (Salon ma'lumotlari, xizmatlar)\n    - 📅 **Vaqt jadvali boshqaruvi** (Booking, Schedule)\n    - 💳 **To'lov tizimi** (Click.uz integratsiyasi)\n    - 📱 **SMS xabarnomalar** (Eskiz.uz integratsiyasi)\n    - 🌐 **Tarjima xizmati** (DeepL API)\n    \n    ### Texnologiyalar:\n    - **FastAPI** - Web framework\n    - **SQLAlchemy** - ORM\n    - **PostgreSQL/SQLite** - Database\n    - **JWT** - Authentication\n    - **Pydantic** - Data validation\n    \n    ### Versiya: 2.0.0\n    ",
    "contact": { "name": "Freya Development Team", "email": "dev@freya.uz" },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "2.0.0"
  },
  "servers": [
    { "url": "http://localhost:8000", "description": "Development server" },
    {
      "url": "https://freya-salon-backend.herokuapp.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/admin/salons": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get All Salons",
        "description": "Barcha salonlar ro'yxatini olish (pagination bilan)",
        "operationId": "get_all_salons_api_admin_salons_get",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Search"
            }
          },
          {
            "name": "is_top",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "boolean" }, { "type": "null" }],
              "title": "Is Top"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "boolean" }, { "type": "null" }],
              "title": "Is Active"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/app__schemas__admin__SalonListResponse"
                  },
                  "title": "Response Get All Salons Api Admin Salons Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/employees/salon/{salon_id}": {
      "get": {
        "tags": ["employees"],
        "summary": "Get Employees By Salon Id",
        "description": "Get employees by salon ID",
        "operationId": "get_employees_by_salon_id_api_employees_salon__salon_id__get",
        "parameters": [
          {
            "name": "salon_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Salon Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/api/salons/": {
      "post": {
        "tags": ["Salons"],
        "summary": "Create Salon",
        "description": "Yangi salon yaratish",
        "operationId": "create_salon_api_salons__post",
        "security": [{ "HTTPBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SalonCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StandardResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Salons"],
        "summary": "Get All Salons",
        "description": "Barcha salonlarni olish",
        "operationId": "get_all_salons_api_salons__get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Search"
            }
          },
          {
            "name": "is_private",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "default": "",
              "title": "Is Private"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__schemas__salon__SalonListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/salons/{salon_id}": {
      "get": {
        "tags": ["Salons"],
        "summary": "Get Salon By Id",
        "description": "ID bo'yicha salonni olish",
        "operationId": "get_salon_by_id_api_salons__salon_id__get",
        "parameters": [
          {
            "name": "salon_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Salon Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SalonResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Salons"],
        "summary": "Update Salon",
        "description": "Salonni yangilash",
        "operationId": "update_salon_api_salons__salon_id__put",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "salon_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Salon Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SalonUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StandardResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Salons"],
        "summary": "Delete Salon",
        "description": "Salonni o'chirish (soft delete)",
        "operationId": "delete_salon_api_salons__salon_id__delete",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "salon_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Salon Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StandardResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/salons/{salon_id}/comments": {
      "post": {
        "tags": ["Salons"],
        "summary": "Add Salon Comment",
        "description": "Salon uchun izoh qo'shish",
        "operationId": "add_salon_comment_api_salons__salon_id__comments_post",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "salon_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Salon Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SalonCommentCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/StandardResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmployeeListResponse": {
        "properties": {
          "success": { "type": "boolean", "title": "Success" },
          "data": {
            "items": { "$ref": "#/components/schemas/EmployeeResponse" },
            "type": "array",
            "title": "Data"
          },
          "pagination": { "type": "object", "title": "Pagination" }
        },
        "type": "object",
        "required": ["success", "data", "pagination"],
        "title": "EmployeeListResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": { "HTTPBearer": { "type": "http", "scheme": "bearer" } }
  }
}